feat(camperData became Map): from a List
I have changed the camperData List to a Map with properties camperList(Map) camperList.allTime(List) camperList.recent(List) and ordering for alltime recent and username. The reason to have the ordering flags to for the toggle headers which will order by ascending or descending. Also changed request to take the period param

closes #4
closes #5