feat(Togglers): allTime and recent
The table headings will toggle which list to show and in what order. I need to add in code that will
change the opposite period in the handleOrder function to empty string again so that when there is
the click back to the other list it will show in descending order by default.

closes #3