refactor(Reducer): camperData functions
The functions that are helping the reducer manage the state are now in a different file. Plus I
fix/changed the order function so that the other period is getting reset. This makes it so when the
user clicks back to the opposite period it will order the list in the default descendin order.

closes #6